% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_column_metadata}
\alias{add_column_metadata}
\title{Add metadata to columns of a data frame}
\usage{
add_column_metadata(.data, ..., .root = list())
}
\arguments{
\item{.data}{A \code{data.frame} or \code{tbl}}

\item{...}{List of metadata in the form \code{column = list(attr1 = value1, attr2 = value2)}}

\item{.root}{List of metadata fields to be applied to data frame itself}
}
\value{
Data, as a \link[sticky:sticky]{sticky} object with metadata stored in \code{attributes}
}
\description{
Add metadata to columns of a data frame
}
