% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata to an object}
\usage{
add_metadata(.data, ...)
}
\arguments{
\item{.data}{Object to which to add metadata}

\item{...}{List of metadata in the form \code{attr1 = value1, attr2 = value2}}
}
\value{
Data as a \link[sticky:sticky]{sticky} object, with metadata stored in \code{attributes}
}
\description{
Add metadata to an object
}
