% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aduToPhotons.R
\name{aduToPhotons}
\alias{aduToPhotons}
\title{Return photons from ADU}
\usage{
aduToPhotons(EMGain)
}
\arguments{
\item{EMGain}{EM gain from the camera}
}
\description{
Compute signal in photons from camera ADU. Set for Hamamatsu cam ImagEM.
}
\author{
JuG
}
