% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_type_dict.R
\docType{data}
\name{base_to_readr_class_dict}
\alias{base_to_readr_class_dict}
\title{Translations from readr classes to base R classes}
\format{An object of class \code{character} of length 6.}
\usage{
base_to_readr_class_dict
}
\description{
Translations from readr classes to base R classes
}
\keyword{datasets}
