\name{yield}
\alias{blastReportStreamer}
\alias{yield}
\title{Streaming records from a \code{\linkS4class{blastReportDB}}.}
\usage{
  blastReportStreamer(con, n = 100)
}
\arguments{
  \item{con}{A \code{blastReportDB} connection.}

  \item{n}{Number of queries to stream}
}
\description{
  Streaming records from a
  \code{\linkS4class{blastReportDB}}.
}

