% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{create_package2}
\alias{create_package2}
\alias{create_package2_sub}
\title{Set up package}
\usage{
create_package2(name, fields = NULL)

create_package2_sub(name)
}
\arguments{
\item{name}{\code{character[1]} giving the package name and path}

\item{fields}{A named list of fields to add to \code{DESCRIPTION}, potentially
overriding default values. See \code{\link[usethis:use_description]{use_description()}} for how you can set
personalized defaults using package options}
}
\value{
Path to the newly created project or package, invisibly.
}
\description{
These functions create an R project:
\itemize{
\item \code{create_package()} creates an R package
\item \code{create_project()} creates a non-package project, i.e. a data analysis
project
}

Both functions can be called on an existing project; you will be asked
before any existing files are changed.
}
