\name{createKronaFile}
\alias{createKronaFile}
\title{create an input file for Krona Webtools}
\usage{
createKronaFile(taxonomyReportDB, output, taxon_db)
}
\arguments{
  \item{taxonomyReportDB}{\code{taxonomyReportDB} object}

  \item{output}{location and name of the output file}
}
\description{
creates an tab-separated input file for the text import
funtion of the Krona Webtools consisting of the the count
of taxons and the corresponding linage
}
\seealso{
runKronaWebtools
}

