% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{ejd_pal}
\alias{ejd_pal}
\title{EJD palette}
\usage{
ejd_pal()
}
\value{
A tibble with three colums, all as character vectors of hex color values:
\enumerate{
\item \code{name} (not a "real" color name),
\item the \code{dark} variant and
\item the \code{light} variant.
}
}
\description{
Color pallete by \href{https://twitter.com/emilyjanedennis}{Dr. Emily Jane Dennis} as per \href{https://twitter.com/debivort/status/994583058031546369}{this tweet}.
}
\details{
Should be color blind friendly
}
\examples{
purrr::walk(.x = ejd_pal()[-1], .f = scales::show_col)

}
\seealso{
Other corporate identity: 
\code{\link{subugoe_template}},
\code{\link{ugoe_pal}()}
}
\concept{corporate identity}
