% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csvy.R
\name{extract_attributes}
\alias{extract_attributes}
\title{Convert metadata list to attributes list suitable for \link{add_metadata}}
\usage{
extract_attributes(meta_data)
}
\arguments{
\item{meta_data}{Named list of metadata values returned by
\link{read_yaml_header}}
}
\description{
Convert metadata list to attributes list suitable for \link{add_metadata}
}
