% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fill_time_gaps}
\alias{fill_time_gaps}
\title{Fill time gaps}
\usage{
fill_time_gaps(x)
}
\arguments{
\item{x}{time vector}
}
\description{
Fill time gaps
}
\author{
M. Saenger, Zimmerberg Risk Analytics GmbH
}
