% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csvy.R
\name{get_all_metadata}
\alias{get_all_metadata}
\title{Extract all metadata from a data.frame and format for YAML header}
\usage{
get_all_metadata(.data)
}
\arguments{
\item{.data}{Data frame from which to extract metadata}
}
\description{
Extract all metadata from a data.frame and format for YAML header
}
