% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_metar_stn}
\alias{get_metar_stn}
\title{Look up airport information}
\usage{
get_metar_stn(
  id.icao = metar.stn$icao,
  fi.name = ".+",
  fi.icao = ".+",
  fi.ctry = ".+",
  fi.lat = c(-90, 90),
  fi.lon = c(-180, 180),
  is.active = TRUE
)
}
\arguments{
\item{id.icao}{Vector of ICAO identifiers}

\item{fi.name}{Lorem Ipsum}

\item{fi.icao}{Lorem Ipsum}

\item{fi.ctry}{Lorem Ipsum}

\item{fi.lat}{Lorem Ipsum}

\item{fi.lon}{Lorem Ipsum}

\item{is.active}{Filter by active stations}
}
\description{
Look up airport information
}
\examples{
# By ICAO Identifier
get_metar_stn(id.icao = c("LSZH", "YSSY"))
# Central Europe
get_metar_stn(fi.icao = "^L", fi.lon = c(5, 11), fi.lat = c(46, 49))
# South America (Argentina)
get_metar_stn(fi.icao = "^SA")
# Contry filter (New Zealand)
get_metar_stn(fi.ctry = "Brazil", fi.lat = c(-40, -30))

}
\author{
M. Saenger
}
