% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_range}
\alias{get_range}
\title{Determine value range}
\usage{
get_range(x, para)
}
\arguments{
\item{x}{vector}

\item{para}{parameter}
}
\description{
Determine value range
}
\author{
M. Saenger, Zimmerberg Risk Analytics GmbH
}
