% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csvy.R
\name{get_readr_col_types}
\alias{get_readr_col_types}
\title{Get the column specification of an existing data frame}
\usage{
get_readr_col_types(.data)
}
\arguments{
\item{.data}{Any \code{data.frame} or \code{tbl_df}}
}
\value{
Column specification, just like \link[readr:spec]{readr::spec}
}
\description{
Effectively, does the same thing as \link[readr:spec]{readr::spec}, but works on
in-memory data frames.
}
\details{
See also \link[readr:spec]{readr::spec} and \link[readr:cols]{readr::cols}.
}
