% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAcqTime.R
\name{getAcqTime}
\alias{getAcqTime}
\title{Get Acquisition time  - POSIXct object}
\usage{
getAcqTime(path)
}
\arguments{
\item{path}{Path to the .txt or .gz metadata file}
}
\value{
Acquisition time (POSIXct object)
}
\description{
Get Acquisition time
}
\author{
JuG
}
