% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDcamExposure.R
\name{getDcamExposure}
\alias{getDcamExposure}
\title{Get DCAM Exposure}
\usage{
getDcamExposure(path)
}
\arguments{
\item{path}{Path to the .txt or .gz metadata file}
}
\value{
DCAM exposure in ms
}
\description{
Get DCAM Exposure - expressed in msec
}
\author{
JuG
}
