% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getElapsedTime.R
\name{getElapsedTime}
\alias{getElapsedTime}
\title{Function to extract elapsed time (in msec)}
\usage{
getElapsedTime(path)
}
\arguments{
\item{path}{Path to the .txt or .gz metadata file}
}
\value{
Elapsed time (vector) in msec
}
\description{
Get elapsed time in ms between consecutive frames
}
\author{
JuG
}
