% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmGain.R
\name{getEmGain}
\alias{getEmGain}
\title{Get EM Gain}
\usage{
getEmGain(path)
}
\arguments{
\item{path}{Path to the .txt or .gz metadata file}
}
\value{
EMGain value
}
\description{
Get EM Gain
}
\author{
JuG
}
