% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExposure.R
\name{getExposure}
\alias{getExposure}
\title{Get Exposure time in ms}
\usage{
getExposure(path)
}
\arguments{
\item{path}{Path to the .txt or .gz metadata file}
}
\value{
Exposure time in ms
}
\description{
Get Exposure time in ms
}
\author{
JuG
}
