% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFrameNumber.R
\name{getFrameNumber}
\alias{getFrameNumber}
\title{Get the total number of frames}
\usage{
getFrameNumber(path)
}
\arguments{
\item{path}{Path to the .txt or .gz metadata file}
}
\value{
EMGain value
}
\description{
Get frame number
}
\author{
JuG
}
