% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImageHeight.R
\name{getImageHeight}
\alias{getImageHeight}
\title{Get Image Height}
\usage{
getImageHeight(path)
}
\arguments{
\item{path}{Path to the .txt or .gz metadata file}
}
\value{
Image height in px
}
\description{
Get image height - expressed in pixels
}
\author{
JuG
}
