% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImageWidth.R
\name{getImageWidth}
\alias{getImageWidth}
\title{Get Image Width}
\usage{
getImageWidth(path)
}
\arguments{
\item{path}{Path to the .txt or .gz metadata file}
}
\value{
Image width in px
}
\description{
Get image width - expressed in pixels
}
\author{
JuG
}
