% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{getMetaGeneric}
\alias{getMetaGeneric}
\title{Generic function to get a given parameter from the metadata file}
\usage{
getMetaGeneric(path, parameter = "FrameIndex", type = "ai")
}
\arguments{
\item{path}{Path to the .txt or .gz metadata file}

\item{parameter}{parameter to extract (string)}

\item{type}{i:integer, n:numeric, f:factor, t:time, ai:'asis'}
}
\value{
Values of the parameter for each image
}
\description{
Generic function to get a given parameter from the metadata file    need to define time zone using Sys.setenv(TZ='GMT') - otherwise warning message
}
\examples{
pNames <- getParamNames(path)
pNames[59]
getMetaGeneric(path, parameter = pNames[59], type = "ai")

getMetaGeneric(path, parameter = "Time", type = "t")
}
\author{
JuG
}
