% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamNames.R
\name{getParamNames}
\alias{getParamNames}
\title{Get Metadata parameters names}
\usage{
getParamNames(path)
}
\arguments{
\item{path}{Path to the .txt or .gz metadata file}
}
\value{
Metadata parameters names
}
\description{
Get metadata parameters names
}
\author{
JuG
}
