% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXpos.R
\name{getXpos}
\alias{getXpos}
\title{Function to extract X position in micrometer}
\usage{
getXpos(path)
}
\arguments{
\item{path}{Path to the .txt or .gz metadata file}
}
\value{
X position (vector) of images
}
\description{
Get X position in micrometer
}
\author{
JuG
}
