% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYpos.R
\name{getYpos}
\alias{getYpos}
\title{Function to extract Y position in micrometer}
\usage{
getYpos(path)
}
\arguments{
\item{path}{Path to the .txt or .gz metadata file}
}
\value{
Y position (vector) of images
}
\description{
Get Y position in micrometer
}
\author{
JuG
}
