% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZpos.R
\name{getZpos}
\alias{getZpos}
\title{Get Z position in micrometer}
\usage{
getZpos(path)
}
\arguments{
\item{path}{Path to the .txt or .gz metadata file}
}
\value{
Z position (vector) of images
}
\description{
Get Z position in micrometer
}
\author{
JuG
}
