\docType{methods}
\name{has_ranks}
\alias{has_ranks}
\title{Check ranks of taxa against a reference}
\usage{
has_ranks(x, ranks, ...)
}
\arguments{
  \item{x}{taxon object}

  \item{ranks}{character vector of taxonomic ranks}
}
\value{
A logical.
}
\description{
Check ranks of taxa against a reference
}

