\name{KronaTable}
\alias{KronaTable}
\title{classify tax_id(s) by taxRanks}
\usage{
  KronaTable()
}
\arguments{
  \item{df}{data.frame for classification}

  \item{taxRank}{valid rank from ncbi taxonomy}

  \item{taxon_db}{connection object to taxonomy db}
}
\value{
  x
}
\description{
  classify tax_id(s) by taxRanks
}

