% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2df.R
\name{list2df}
\alias{list2df}
\title{Convert list to data.frame with list columns where appropriate}
\usage{
list2df(l)
}
\arguments{
\item{l}{List to convert to a data.frame}
}
\value{
\code{tbl} form of the list, with list columns where appropriate
}
\description{
Convert list to data.frame with list columns where appropriate
}
