\docType{methods}
\name{metadata}
\alias{Access metadata}
\alias{metadata}
\alias{metadata,taxonomyReportDB-method}
\alias{metadata<-,taxonomyReportDB-method}
\title{Access metadata}
\usage{
metadata(x, ...)

\S4method{metadata}{taxonomyReportDB}(x)

metadata(x, ...) <- value
}
\description{
Access metadata
}

