% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{metar.class}
\alias{metar.class}
\title{METAR Parameter Classes}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 54 rows and 10 columns.
}
\usage{
metar.class
}
\description{
METAR Parameter Classes
}
\keyword{datasets}
