% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{metar.dc}
\alias{metar.dc}
\title{Present Weather Descriptor}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 8 rows and 2 columns.
}
\usage{
metar.dc
}
\description{
Present Weather Descriptor
}
\keyword{datasets}
