% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{metar.int}
\alias{metar.int}
\title{Present Weather Intensity}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 3 rows and 3 columns.
}
\usage{
metar.int
}
\description{
Present Weather Intensity
}
\keyword{datasets}
