% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{metar.para}
\alias{metar.para}
\title{METAR Parameters}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 62 rows and 9 columns.
}
\usage{
metar.para
}
\description{
METAR Parameters
}
\keyword{datasets}
