% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{metar.para_conv}
\alias{metar.para_conv}
\title{METAR Parameter Conversion}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 12 rows and 6 columns.
}
\usage{
metar.para_conv
}
\description{
METAR Parameter Conversion
}
\keyword{datasets}
