% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{metar.ph}
\alias{metar.ph}
\title{Present Weather Phenomena}
\format{
An object of class \code{list} of length 23.
}
\usage{
metar.ph
}
\description{
Present Weather Phenomena
}
\keyword{dataset}
