% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{metar.ph_grp}
\alias{metar.ph_grp}
\title{Present Weather Phenomena Group}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 6 rows and 2 columns.
}
\usage{
metar.ph_grp
}
\description{
Present Weather Phenomena Group
}
\keyword{datasets}
