% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{metar.ph_list}
\alias{metar.ph_list}
\title{Present Weather Phenomena List}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 402 rows and 2 columns.
}
\usage{
metar.ph_list
}
\description{
Present Weather Phenomena List
}
\keyword{datasets}
