% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{metar.ph_subgrp}
\alias{metar.ph_subgrp}
\title{Present Weather Phenomena Sub-Group}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 8 rows and 2 columns.
}
\usage{
metar.ph_subgrp
}
\description{
Present Weather Phenomena Sub-Group
}
\keyword{datasets}
