% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{metar.stn}
\alias{metar.stn}
\title{Station meta data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 22296 rows and 15 columns.
}
\usage{
metar.stn
}
\description{
Station meta data
}
\author{
http://www.weathergraphics.com/identifiers/
}
\keyword{datasets}
