% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{metar.test}
\alias{metar.test}
\title{METAR Test Data Set}
\format{
An object of class \code{character} of length 21.
}
\usage{
metar.test
}
\description{
METAR Test Data Set
}
\keyword{dataset}
