% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{metar.vars.cld}
\alias{metar.vars.cld}
\title{Cloud Output Variables}
\format{
An object of class \code{list} of length 4.
}
\usage{
metar.vars.cld
}
\description{
Cloud Output Variables
}
\keyword{dataset}
