% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{metar.vars.pw}
\alias{metar.vars.pw}
\title{Present Weather Output Variables}
\format{
An object of class \code{list} of length 4.
}
\usage{
metar.vars.pw
}
\description{
Present Weather Output Variables
}
\keyword{dataset}
