% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{metar.vars}
\alias{metar.vars}
\title{METAR Output Variables}
\format{
An object of class \code{list} of length 32.
}
\usage{
metar.vars
}
\description{
METAR Output Variables
}
\keyword{dataset}
