% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{metar_latest}
\alias{metar_latest}
\title{Get METARs from NOAA cycle files}
\usage{
metar_latest(id_icao = "", latest.only = TRUE, report.hour = NULL)
}
\arguments{
\item{id_icao}{ICAO identifiery, e. g. YSSY}

\item{latest.only}{latest report per aiport only}

\item{report.hour}{specific reporting hour (last 24h)}
}
\description{
metar_latest(id_icao = "LSZH")
metar_latest(id_icao = "EG")
}
\author{
M. Saenger
}
