% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{metar_pw}
\alias{metar_pw}
\title{Parse METAR PW groups}
\usage{
metar_pw(pw, DC = FALSE, PH = FALSE)
}
\arguments{
\item{pw}{METAR PW string}

\item{DC}{Add descriptor cross table}

\item{PH}{Add phenomenon cross table}
}
\description{
Parse METAR PW groups
}
\author{
M. Saenger
}
