% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{metar_rvr}
\alias{metar_rvr}
\title{Parse METAR RVR groups}
\usage{
metar_rvr(rvr)
}
\arguments{
\item{rvr}{METAR RVR string}
}
\description{
Parse METAR RVR groups
}
\examples{
metar_rvr(rvr = "R14/P2000N R16/P2000N R28/P2000N R34/1000VP2000U")

}
\author{
M. Saenger
}
