% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{metar_validate}
\alias{metar_validate}
\title{Validate METAR}
\usage{
metar_validate(dat, set.na = TRUE)
}
\arguments{
\item{dat}{Output from \code{metar_parse}}

\item{set.na}{set NA if validation fails}
}
\description{
Validate METAR reports
}
\author{
M. Saenger
}
