% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{multicolor.title}
\alias{multicolor.title}
\title{Plot METARgram}
\usage{
multicolor.title(main, col.main, collapse = "", ...)
}
\arguments{
\item{main}{text vector}

\item{col.main}{corresponding colour vector}

\item{collapse}{collapse}

\item{...}{further arguments for title()}
}
\description{
Plot METARgram
}
\author{
https://github.com/CollinErickson/ContourFunctions/blob/master/R/cf_multicolor_title.R
}
