% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_metar}
\alias{parse_metar}
\title{Parse METAR Groups}
\usage{
parse_metar(x, t = NULL, yyyy = year(Sys.Date()), mm = month(Sys.Date()))
}
\arguments{
\item{x}{METAR reports (character vector)}

\item{t}{time (POSIXct vector of same length as \code{x})}

\item{yyyy}{year}

\item{mm}{month}
}
\description{
Parse a METAR report into groups
Performance: ~ 60s for 1 million records on average machine
}
\examples{
x <- metar.test$code
parse_metar(x)
parse_metar(x = "EGVP 281509Z 27011KT 5000 1000SW +SHRA BKN022CB 15/13 Q1007 BECMG 9999 NSW")

}
\author{
M. Saenger, Zimmerberg Risk Analytics GmbH
}
