% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{parse_metar_pw}
\alias{parse_metar_pw}
\title{Parse present wather groups}
\usage{
parse_metar_pw(x)
}
\arguments{
\item{x}{present weather string}
}
\description{
Parse present wather groups
}
\examples{
x <- c("TS", "+TSRA VCTS", "SNGRRA VCFG")
parse_metar_pw(x)
m <- parse_metar(metar.test$code)
parse_metar_pw(x = m$pw)

}
\author{
M. Saenger, Zimmerberg Risk Analytics GmbH
}
