% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{parse_metar_rvr}
\alias{parse_metar_rvr}
\title{Parse runway visibility groups}
\usage{
parse_metar_rvr(x)
}
\arguments{
\item{x}{cloud string}
}
\description{
Parse runway visibility groups
}
\examples{
x <- c("R14/P2000N R16/P2000N R28/P2000N R34/1000VP2000U", "R08R/3000VP6000FT", "R22/0700D", "R01/0900VP2000D")
parse_metar_rvr(x)

}
\author{
M. Saenger
}
