% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{parse_metar_rwy_cond}
\alias{parse_metar_rwy_cond}
\title{Parse runway conditions groups}
\usage{
parse_metar_rwy_cond(x)
}
\arguments{
\item{x}{cloud string}
}
\description{
Parse runway conditions groups
}
\examples{
x <- c("R02/010070 R06/010070", "R88/090060")
parse_metar_rwy_cond(x)

}
\author{
M. Saenger, Zimmerberg Risk Analytics GmbH
}
