% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{parse_metar_ws}
\alias{parse_metar_ws}
\title{Parse wind shear}
\usage{
parse_metar_ws(x)
}
\arguments{
\item{x}{cloud string}
}
\description{
Parse wind shear
}
\examples{
x <- c("WS R25R")
parse_metar_ws(x)

}
\author{
M. Saenger, Zimmerberg Risk Analytics GmbH
}
