% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_metar_wx}
\alias{parse_metar_wx}
\title{Parse METAR Message}
\usage{
parse_metar_wx(x)
}
\arguments{
\item{x}{METAR reports (character vector)}
}
\description{
Parse a METAR current weather section
}
\examples{
x <- parse_metar(metar.test$code)$wx
parse_metar_wx(x)

}
\author{
M. Saenger
}
