% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_metargram}
\alias{plot_metargram}
\title{Plot METARgram}
\usage{
plot_metargram(
  dat,
  cex = 0.9,
  attribution = "Data: Iowa State Univ. mesonet.agron.iastate.edu/"
)
}
\arguments{
\item{dat}{data set}

\item{cex}{global text size}

\item{attribution}{data source}
}
\description{
Plot METARgram
}
\author{
M. Saenger
}
