% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar.R
\name{process_metar}
\alias{process_metar}
\title{Convert \code{data.frame} to \code{list}.}
\usage{
process_metar(dat, month = NULL, year = NULL, col.drop = NULL,
  verbose = F)
}
\arguments{
\item{dat}{A \code{data.frame} object.}

\item{month}{Set month of METAR report}

\item{year}{Set year of METAR report}

\item{col.drop}{Columns to drop after processing}

\item{verbose}{Show comments}
}
\description{
processes a data.frame of parsed METAR reports
}
\examples{
process_metar(parse_metar("LSZH 271750Z 14002KT CAVOK 25/18 Q1017 NOSIG"))
}
\author{
M. Saenger
}
