% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csvy.R
\name{read_csvy}
\alias{read_csvy}
\title{Read a CSVY file}
\usage{
read_csvy(file, metadata = NULL, verbose = TRUE, ...)
}
\arguments{
\item{file}{Name of file to read.}

\item{metadata}{Path to additional metadata. If \code{NULL} (default), assume
metadata is in \code{file}.}

\item{verbose}{Logical. If \code{TRUE}, print warning if no header found.}

\item{...}{Additional arguments to \link[readr:read_csv]{readr::read_csv}}
}
\description{
Read a CSVY file
}
\details{
A CSVY (CSV + YAML) file is just a CSV file prepended with a YAML
header containing relevant metadata. The metadata can be any list or
list-like object, and any fields not specially recognized will be parsed by
\link[yaml:read_yaml]{yaml::read_yaml} and stored as named attributes. The following metadata
fields are parsed specially:
\itemize{
\item \code{read_csv} -- contains the arguments to be passed to \link[readr:read_csv]{readr::read_csv},
such as the field separator (\code{sep}) and \code{NA} strings (\code{na.strings}). In
general, this is not necessary because \code{read_csv} has sensible defaults and
is clever about figuring such things out, but it may improve performance or
reliability for particularly large or complex files.
\item \code{resources: fields} (Single item \code{fields} nested inside root item
\code{resources}) -- A list of metadata for the table columns.  These are
structured generally following the \href{https://frictionlessdata.io/specs/table-schema}{TableSchema} of a \href{https://frictionlessdata.io/specs/tabular-data-package}{Tabular DataPackage} with a few
R-friendly modifications. All of these fields are optional, though some can
be deduced automatically during the writing process. Some common fields are
as follows (fields with \code{*} following their labels are not in the Table
Schema):
\itemize{
\item \code{name} -- Name of field, corresponding to column name in data
\item \code{title} -- A human-readable label for the field
\item \code{description} -- A description for the field
\item \code{type} -- String specifying the schema type (see \link{schema_type_dict})
\item \code{class*} -- String or array specifying the R \code{class}(es) to which the
object belongs.
\item \code{constraints} -- A list of constraints on the data (see \href{https://frictionlessdata.io/specs/table-schema/#constraints}{Table Schema:Constraints}).
Eventually, all of these will be actively used by R to verify data.
}
\item \code{missingValues} -- An array of strings that are used to indicate missing
values.
}
}
