% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_fct.R
\name{read_mesonet}
\alias{read_mesonet}
\title{Read METAR reports from mesonet.agron.iastate.edu}
\usage{
read_mesonet(
  id_icao = "HSSS",
  date_start = date_end - 3600 * 12,
  date_end = Sys.time(),
  auto = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{id_icao}{Lorem Ipsum}

\item{date_start}{Lorem Ipsum}

\item{date_end}{Lorem Ipsum}

\item{auto}{Lorem Ipsum}

\item{verbose}{Lorem Ipsum}
}
\description{
Read METAR reports from mesonet.agron.iastate.edu
}
\examples{
read_mesonet(id_icao = "KDEN", date_start = Sys.time() - 3600*24*1)

}
\author{
M. Saenger
}
