% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_metar_noaa}
\alias{read_metar_noaa}
\title{Get METARs from NOAA cycle files}
\usage{
read_metar_noaa(hour, remote = TRUE, path, latest.only = TRUE)
}
\arguments{
\item{hour}{specific reporting hour (last 24h)}

\item{remote}{Download file from NOAA server}

\item{path}{Path to local file}

\item{latest.only}{latest report per aiport only}
}
\description{
read_metar_noaa(hour = 15, latest.only = T)
}
\author{
M. Saenger
}
