% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_fct.R
\name{read_station}
\alias{read_station}
\title{Get airport information from https://www.aviationweather.gov}
\usage{
read_station(
  fi.name = ".+",
  fi.icao = ".+",
  fi.ctry = ".+",
  fi.lat = c(-90, 90),
  fi.lon = c(-180, 180)
)
}
\arguments{
\item{fi.name}{Lorem Ipsum}

\item{fi.icao}{Lorem Ipsum}

\item{fi.ctry}{Lorem Ipsum}

\item{fi.lat}{Lorem Ipsum}

\item{fi.lon}{Lorem Ipsum}
}
\description{
Get airport information from https://www.aviationweather.gov
}
\examples{
read_station(fi.icao = "^L", fi.lon = c(5, 11), fi.lat = c(46, 49))

}
\author{
M. Saenger
}
