% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_yaml_header.R
\name{read_yaml_header}
\alias{read_yaml_header}
\title{Retrieve YAML header from file}
\usage{
read_yaml_header(file, verbose = TRUE, ...)
}
\arguments{
\item{file}{Filename from which to read YAML header}

\item{verbose}{Logical. If \code{TRUE}, print warning if no header found.}

\item{...}{Additional arguments to \link[yaml:read_yaml]{yaml::read_yaml}}
}
\value{
Character vector of lines containing YAML header, or \code{NULL} if no
YAML header found.
}
\description{
Note that this assumes only one Yaml header, starting on the first line of
the file.
}
