% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pkgdown_template}
\alias{ugoe_pal}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{subugoetheme}{\code{\link[subugoetheme]{pkgdown_template}}, \code{\link[subugoetheme]{ugoe_pal}}}
}}

