% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_type_dict.R
\docType{data}
\name{schema_type_dict}
\alias{schema_type_dict}
\title{Translations between Table Schema types and readr column specifications.}
\format{An object of class \code{character} of length 7.}
\usage{
schema_type_dict
}
\description{
Translations between Table Schema types and readr column specifications.
}
\keyword{datasets}
