% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{set_alpha}
\alias{set_alpha}
\title{Set Alpha}
\usage{
set_alpha(x, alpha = 1)
}
\arguments{
\item{x}{Colour vector}

\item{alpha}{Transparency}
}
\description{
Set Alpha
}
\examples{
set_alpha(x = c("red", "blue"), alpha = c(0.1, .3))

}
