% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{subugoe_template}
\alias{subugoe_template}
\title{State- and University Library of Göttingen \href{http://pkgdown.r-lib.org}{pkgdown} template}
\usage{
subugoe_template()
}
\description{
A list of paths to assets and templates for web design.
}
\details{
To use this template for website, list {metar} in your \code{DESCRIPTION} and add to your \verb{_pkgdown.yml}:\if{html}{\out{<div class="yaml">}}\preformatted{template:
  package: metar
}\if{html}{\out{</div>}}

As a result, your website will:
\itemize{
\item Use the \href{https://bootswatch.com/3/cosmo/}{bootswatch \code{cosmo} bootstrap theme}.
\item Set the navbar to \code{type: inverse}.
\item Display in a University of Göttingen \strong{color scheme}.
This is determined by the \code{ugoe_bootswatch.css}, which is compiled from \code{ugoe_colors.less}.
The colors were added and downloaded via https://www.bootstrap-live-customizer.com.
\item Include the SUB wordmark in your website as \code{sub_wordmark.jpg}.
\item Add a custom footer including the wordmar.
}

To pull in changes in this package to your website, install the current version of {metar} and re-run \code{pkgdown::build_site()} locally or in your CI.
To pull in upstream changes from {pkgdown} into \emph{this} package, it may be ocassionally necessary to update the files in \code{inst/pkgdown/templates} which have been minimally adapted from their original counterparts in {pkgdown}.

To learn more about pkgdown templates, read \href{https://pkgdown.r-lib.org/reference/build_site.html}{here} or consider \href{https://github.com/ropensci/rotemplate}{these} \href{https://tidytemplate.tidyverse.org}{examples}.
}
\seealso{
Other corporate identity: 
\code{\link{ejd_pal}()},
\code{\link{ugoe_pal}()}
}
\concept{corporate identity}
