% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{ugoe_pal}
\alias{ugoe_pal}
\title{University of Göttingen palette}
\usage{
ugoe_pal()
}
\value{
A list of three elements, all character vectors of hex color values, named by their official name as per the above source:
\enumerate{
\item \code{primary} colors,
\item \code{secondary} colors and
\item \code{faculty} for the colors of the different schools.
}
}
\description{
Color pallete used at the \href{https://www.uni-goettingen.de}{Georg August University of Göttingen} as per \href{https://www.uni-goettingen.de/de/589412.html}{these guidelines}.
}
\examples{
purrr::walk(.x = ugoe_pal(), .f = scales::show_col)

}
\seealso{
Other corporate identity: 
\code{\link{ejd_pal}()}
}
\concept{corporate identity}
