% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{use_ghpages}
\alias{use_ghpages}
\title{Set up GitHub Pages}
\usage{
use_ghpages(name)
}
\arguments{
\item{name}{\verb{[character(1)]} giving the package name and path}
}
\description{
Set up GitHub Pages
}
