% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{use_github_actions2}
\alias{use_github_actions2}
\title{Set up github actions}
\usage{
use_github_actions2()
}
\description{
Set up github actions
}
\seealso{
Other setup helpers: 
\code{\link{create_package2}()},
\code{\link{use_codecov2}()},
\code{\link{use_ghpages}()},
\code{\link{use_pkgdown2}()}
}
\concept{setup helpers}
