% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_labels.R
\name{use_wag_labels}
\alias{use_wag_labels}
\alias{wag_labels}
\title{Set WAG default issue labels}
\usage{
use_wag_labels()

wag_labels()
}
\description{
Wraps to \code{\link[usethis:use_github_labels]{usethis::use_github_labels()}} to add WAG default labels to GitHub defaults.
}
\examples{
# these labels are added
wag_labels()

# just run this to set up labels in your current repo
\dontrun{
use_wag_labels()
}
}
