% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{validate_metar}
\alias{validate_metar}
\title{Validate METAR}
\usage{
validate_metar(dat, formulas, verbose = TRUE)
}
\arguments{
\item{dat}{Output from \code{parse_metar}}

\item{formulas}{collection of validation formulas as quosures e.g. \code{rlang::quos(f1 = tt > tt, f2 = ff > fx)}}

\item{verbose}{verbose output}
}
\description{
Validate METAR reports
}
\author{
M. Saenger, Zimmerberg Risk Analytics GmbH
}
