% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csvy.R
\name{write_csvy}
\alias{write_csvy}
\title{Write data and its metadata to a CSVY file}
\usage{
write_csvy(.data, filename, comment = "#", ...)
}
\arguments{
\item{.data}{\code{data.frame} with metadata}

\item{filename}{Output file name}

\item{comment}{Comment string with which to prefix file. Default = "#".}

\item{...}{Additional arguments to \link[data.table:fwrite]{data.table::fwrite}}
}
\description{
Write data and its metadata to a CSVY file
}
