% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_yaml.R
\name{write_yaml}
\alias{write_yaml}
\title{Write list as YAML file}
\usage{
write_yaml(l, file, comment = "#", ...)
}
\arguments{
\item{l}{List to write out to file}

\item{file}{Path to target file}

\item{comment}{Comment string with which to prefix file. Default = "#".}

\item{...}{Additional arguments to \link[yaml:as.yaml]{yaml::as.yaml}}
}
\value{
\code{NULL}, invisibly
}
\description{
Write list as YAML file
}
