% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{wx_es}
\alias{wx_es}
\title{Vapour pressure saturated (wrapper function)}
\usage{
wx_es(tt, ice = FALSE)
}
\arguments{
\item{tt}{°C}

\item{ice}{true/false}
}
\description{
Vapour pressure saturated (wrapper function)
ECMWF IFS Documentation – Cy41r2, Part IV, formula 7.5
}
\examples{
wx_es(30, TRUE)

}
\author{
M. Saenger
}
