% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\name{wx_es_i}
\alias{wx_es_i}
\title{Vapour pressure saturated (over ice)}
\usage{
wx_es_i(tt)
}
\arguments{
\item{tt}{°C}
}
\description{
ECMWF IFS Documentation – Cy41r2, Part IV, formula 7.5
}
\examples{
wx_es_i(-30)

}
\author{
M. Saenger
}
