% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{wx_es_w}
\alias{wx_es_w}
\title{Vapour pressure saturated (over water)}
\usage{
wx_es_w(tt)
}
\arguments{
\item{tt}{°C}
}
\description{
ECMWF IFS Documentation – Cy41r2, Part IV, formula 7.5
}
\examples{
wx_es_w(30)

}
\author{
M. Saenger
}
