% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{wx_rh}
\alias{wx_rh}
\title{Relative humidity}
\usage{
wx_rh(tt, td, ice = FALSE)
}
\arguments{
\item{tt}{°C Air temperature}

\item{td}{°C Dew point}

\item{ice}{over ice}
}
\description{
Relative humidity
}
\examples{
wx_rh(5, -35)

}
\author{
M. Saenger
}
