% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get meta data}
\usage{
get_metadata(data, example = TRUE)
}
\arguments{
\item{data}{The data you want meta data on - can be in the form of a
Data Frame, Data Table or Tibble (possible others).}

\item{example}{Logical Vector; If TRUE (default) it will add a colum in
the return with an example of an averagely selected (non NA) data point.}
}
\value{
A data vector of type input (if data is Data Table, the returned
value is Data Table, if input is Data Frame the returned value is Data Frame etc.)

The returned data will include the following columns:
\itemize{
  \item \strong{name:} The name of the column
  \item \strong{colNo:} The position of the column, counting fom the left
  \item \strong{class:} The class type of the column
  \item \strong{na_count:} The number of NA values in the column
  \item \strong{null_count:} The number of NULL values in the column
  \item \strong{empty_count:} The number of NA and/or NULL values in the column
  \item \strong{empty_pct:} The percentage of empty values in the column
  \item \strong{unique_count:} The number of unique values in the column
  \item \strong{unique_pct:} The percentage of unique values in the column
  \item \strong{example:} (optional) Examples of data in the column. For numeric data
  selected by mean and for text data, selected by the avg. length of the text.
}
}
\description{
Get meta data on a Data Frame, Data Table, Tibble etc.
}
\details{
This function will return a table with relevant meta data on your input in a nice
tidy table structure including name, class, number of empty cells, percents
empty cells, number of unique values, percents of unique values and examples of
data for each column in the inputtet dataset.
}
\examples{

df <-
  data.frame(
    letters = letters,
    num = 1:26,
    letters2 = c(rep("a", 4),rep(NA_character_, 6), rep("b", 10), rep(NA_character_, 6)),
    date = as.POSIXct("2010-01-01"))

df_meta <- get_metadata(df)

print(df_meta)

}
