\name{getQueryDef}
\alias{getCogID}
\alias{getGeneID}
\alias{getKeggID}
\alias{getQueryDef}
\alias{getScientificName}
\alias{getScore}
\alias{getTaxID}
\title{Getter for metaCVReport}
\arguments{
  \item{metaCVReport}{a \code{metaCVReport} object}

  \item{id}{tax_id}
}
\description{
  Getter for metaCVReport
}

