\docType{methods}
\name{getKeggID}
\alias{getCogID}
\alias{getCogID,metaCVReport-method}
\alias{getGeneID,metaCVReport-method}
\alias{getKeggID}
\alias{getKeggID,metaCVReport-method}
\alias{getQueryDef,metaCVReport-method}
\alias{getScientificName,metaCVReport-method}
\alias{getScore,metaCVReport-method}
\alias{getTaxID,metaCVReport-method}
\title{Access fields of a metaCVReport object.}
\usage{
getKeggID(x, ...)

getCogID(x, ...)

\S4method{getQueryDef}{metaCVReport}(x, id)

\S4method{getScore}{metaCVReport}(x, id)

\S4method{getGeneID}{metaCVReport}(x, id)

\S4method{getKeggID}{metaCVReport}(x, id)

\S4method{getCogID}{metaCVReport}(x, id)

\S4method{getTaxID}{metaCVReport}(x, id)

\S4method{getScientificName}{metaCVReport}(x, id)
}
\arguments{
  \item{x}{a \code{\linkS4class{metaCVReport}} object.}

  \item{...}{}
}
\description{
Access fields of a metaCVReport object.
}

