% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\alias{metadata<-}
\title{Access and set object metadata}
\usage{
metadata(.data, ...)

metadata(.data) <- value
}
\arguments{
\item{.data}{Object to which to add metadata}

\item{...}{Unquoted values to extract from metadata}

\item{value}{List of value-name pairs to be passed to \code{add_metadata}}
}
\description{
Access and set object metadata
}
