% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaR-package.R
\docType{package}
\name{metaR-package}
\alias{metaR}
\alias{metaR-package}
\title{metaR: Internal package for SUB use}
\description{
Wrappers, settings and templates for #rstats work at SUB.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://subugoe.github.io/metaR}
  \item \url{https://github.com/subugoe/metaR}
  \item Report bugs at \url{https://github.com/subugoe/metaR/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Held \email{info@maxheld.de} (\href{https://orcid.org/0000-0002-4703-5388}{ORCID})

Other contributors:
\itemize{
  \item Göttingen State and University Library [copyright holder, funder]
}

}
\keyword{internal}
