% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{metar.test}
\alias{metar.test}
\title{METAR Test Data Set}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 24 rows and 3 columns.
}
\usage{
metar.test
}
\description{
METAR Test Data Set
}
\keyword{datasets}
