% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{metar_clouds}
\alias{metar_clouds}
\title{Parse a METAR cloud groups}
\usage{
metar_clouds(cld)
}
\arguments{
\item{cld}{METAR cloud string}
}
\description{
Parse a METAR cloud groups
}
\examples{
 metar_clouds(cld = c("FEW030 FEW032TCU SCT050 BKN075", "SCT050 SCT075 OVC149", "", "SCT003"))

}
\author{
M. Saenger
}
