% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{metar_latest}
\alias{metar_latest}
\title{Get METARs from NOAA cycle files}
\usage{
metar_latest(id_icao = NULL, latest = TRUE)
}
\arguments{
\item{id_icao}{Lorem Ipsum}

\item{latest}{Lorem Ipsum}
}
\description{
metar_latest(id_icao = "LSZH")
metar_latest(id_icao = "EG")
}
\author{
M. Saenger
}
