% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{parse_metar}
\alias{parse_metar}
\title{Parse a METAR report}
\usage{
parse_metar(x, date = NULL)
}
\arguments{
\item{x}{METAR strings}

\item{date}{date}
}
\description{
Parse a METAR report
}
\examples{
 parse_metar(x = "LSZH 271750Z 14002KT CAVOK 25/18 Q1017 NOSIG")
 parse_metar(x = metar.test)

}
\author{
M. Saenger
}
