% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar.R
\name{parse_metar}
\alias{parse_metar}
\title{Parse a METAR report}
\usage{
parse_metar(str, verbose = FALSE)
}
\arguments{
\item{str}{A METAR string.}

\item{verbose}{Show comments}
}
\description{
Parse a METAR report
}
\examples{
parse_metar("LSZH 271750Z 14002KT CAVOK 25/18 Q1017 NOSIG")
}
\author{
M. Saenger
}
