% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar.R
\name{parse_metar}
\alias{parse_metar}
\title{Convert \code{data.frame} to \code{list}.}
\usage{
parse_metar(str, verbose = F)
}
\arguments{
\item{str}{A \code{character} object.}

\item{verbose}{Show comments}
}
\description{
parse a METAR report
}
\examples{
parse_metar("LSZH 271750Z 14002KT CAVOK 25/18 Q1017 NOSIG")
}
\author{
M. Saenger
}
