% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar.R
\name{process_metar}
\alias{process_metar}
\title{Processes a data.frame of parsed METAR reports}
\usage{
process_metar(dat, month = NULL, year = NULL, col.drop = NULL, verbose = FALSE)
}
\arguments{
\item{dat}{A \code{data.frame} object (output from parse_metar).}

\item{month}{Set month of METAR report}

\item{year}{Set year of METAR report}

\item{col.drop}{Columns to drop after processing}

\item{verbose}{Show comments}
}
\description{
Processes a data.frame of parsed METAR reports
}
\examples{
process_metar(parse_metar("LSZH 271750Z 14002KT CAVOK 25/18 Q1017 NOSIG"))
}
\author{
M. Saenger
}
