% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_metar_mesonet}
\alias{read_metar_mesonet}
\title{Read METAR reports from mesonet.agron.iastate.edu}
\usage{
read_metar_mesonet(
  id_icao = "LSZH",
  remote = TRUE,
  path,
  date_start = date_end - 3600 * 24 * 7,
  date_end = Sys.time() + 3600 * 24,
  verbose = FALSE
)
}
\arguments{
\item{id_icao}{Lorem Ipsum}

\item{remote}{Download file from mesonet.agron.iastate.edu}

\item{path}{Path to local file}

\item{date_start}{Lorem Ipsum}

\item{date_end}{Lorem Ipsum}

\item{verbose}{Lorem Ipsum}
}
\description{
Read METAR reports from mesonet.agron.iastate.edu
read_metar_mesonet(id_icao = "KDEN", date_start = Sys.time() - 3600\emph{24}1, verbose = TRUE)
}
\author{
M. Saenger
}
