\name{getTaxon}
\alias{getAccession}
\alias{getAlignLen}
\alias{getBitscore}
\alias{getDefinition}
\alias{getEvalue}
\alias{getGaps}
\alias{getGeneID}
\alias{getHitFrame}
\alias{getHitFrom}
\alias{getHitID}
\alias{getHitLen}
\alias{getHitNum}
\alias{getHitSeq}
\alias{getHitTo}
\alias{getHspID}
\alias{getHspNum}
\alias{getIdentity}
\alias{getLineage}
\alias{getMatch}
\alias{getOtherName}
\alias{getParentTaxID}
\alias{getPositive}
\alias{getQueryDef}
\alias{getQueryFrame}
\alias{getQueryFrom}
\alias{getQueryID}
\alias{getQueryLen}
\alias{getQuerySeq}
\alias{getQueryTo}
\alias{getRank}
\alias{getScientificName}
\alias{getTaxID}
\alias{getTaxon}
\title{Getters for taxonomyReportDB}
\arguments{
  \item{x}{A \code{taxonomyReportDB} object.}

  \item{id}{An identifier}

  \item{type}{Type of identifier. One of \sQuote{tax_id},
  \sQuote{query_id}, or \sQuote{hit_id}.}
}
\description{
  Getters for taxonomyReportDB
}

