% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{validate_metar}
\alias{validate_metar}
\title{Validate METAR}
\usage{
validate_metar(dat, formulas, set.na = TRUE)
}
\arguments{
\item{dat}{Output from \code{parse_metar}}

\item{formulas}{selection of  formulas as quosures \code{rlang::quos(f1 = tt > tt, f2 = ff > fx)}}

\item{set.na}{set NA if validation fails}
}
\description{
Validate METAR reports
}
\author{
M. Saenger
}
